requirements.txt
certifi==2018.8.13
chardet==3.0.4
idna==2.7
mongoengine==0.15.3
pymongo==3.7.1
PyYAML==3.13
requests==2.19.1
six==1.11.0
urllib3==1.23
